server/pyproject.toml
[project]
name = "planarally-server"
version = "2026.1.0"
description = "Planarally server"
readme = "README.md"
requires-python = ">=3.13.2"

dependencies = [
    "aiohttp[speedups]==3.13.3",
    "aiohttp-security==0.5.0",
    "aiohttp-session==2.12.1",
    "bcrypt==5.0.0",
    "cryptography==46.0.5",
    "peewee==3.19.0",
    "pillow==12.1.1",
    "pymupdf==1.25.4",
    "pydantic[email]==2.12.5",
    "python-engineio==4.13.1",
    "python-socketio==5.16.1",
    "redmail==0.6.0",
    "rtoml==0.13.0",
    "typing-extensions==4.15.0",
    "watchdog==6.0.0",
    # Dungeongen extension
    "dungeongen>=0.1.14",
]

[dependency-groups]
dev = ["pydantic-to-typescript==2.0.0", "ruff==0.15.1"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
ignore = ["E722"]

[tool.ruff.lint.per-file-ignores]
"src/api/models/*" = ["F403"]
